Move test helpers (`HTTPClientStub` and `InMemoryFeedStore`) to separate files
Finished lecture 'Validating Acceptance Criteria with Fast Integration Tests, Composition Root, and Simulating App Launch & State Transitions'
