+++ EssentialApp/EssentialAppTests/Helpers/HTTPClientStub.swift
@@ -1,0 +1,36 @@
+//
+//  HTTPClientStub.swift
+//  EssentialAppTests
+//
+//  Created by Finn Ebeling on 15.04.24.
+//
+
+import Foundation
+import EssentialFeed
+
+class HTTPClientStub: HTTPClient {
+    private class Task: HTTPClientTask {
+        func cancel() {}
+    }
+    
+    private let stub: (URL) -> HTTPClient.Result
+    
+    init(stub: @escaping (URL) -> HTTPClient.Result) {
+        self.stub = stub
+    }
+    
+    func get(fromURL url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
+        completion(stub(url))
+        return Task()
+    }
+}
+
+extension HTTPClientStub {
+    static var offline: HTTPClientStub {
+        .init(stub: { _ in .failure(NSError(domain: "offline", code: 0)) })
+    }
+    
+    static func online(_ stub: @escaping (URL) -> (Data, HTTPURLResponse)) -> HTTPClientStub {
+        .init(stub: { url in .success(stub(url)) })
+    }
+}
