+++ EssentialApp/EssentialAppTests/Helpers/InMemoryFeedStore.swift
@@ -1,0 +1,62 @@
+//
+//  InMemoryFeedStore.swift
+//  EssentialAppTests
+//
+//  Created by Finn Ebeling on 15.04.24.
+//
+
+import Foundation
+import EssentialFeed
+
+class InMemoryFeedStore {
+    private(set) var feedCache: CachedFeed?
+    private var feedImageDataCache: [URL: Data]
+    
+    init(feedCache: CachedFeed? = nil, feedImageDataCache: [URL : Data] = [:]) {
+        self.feedCache = feedCache
+        self.feedImageDataCache = feedImageDataCache
+    }
+    
+}
+
+extension InMemoryFeedStore: FeedStore {
+    
+    func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
+        feedCache = nil
+        completion(.success(()))
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
+        feedCache = CachedFeed(feed: feed, timestamp: timestamp)
+        completion(.success(()))
+    }
+    
+    func retrieve(completion: @escaping FeedStore.RetrievalCompletion) {
+        completion(.success(feedCache))
+    }
+}
+
+extension InMemoryFeedStore: FeedImageDataStore {
+    func insert(_ data: Data, for url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
+        feedImageDataCache[url] = data
+        completion(.success(()))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        completion(.success(feedImageDataCache[url]))
+    }
+}
+
+extension InMemoryFeedStore {
+    static var empty: InMemoryFeedStore {
+        InMemoryFeedStore()
+    }
+    
+    static var withExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
+    }
+    
+    static var withNonExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
+    }
+}
